requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.0.4
idna==2.10
newsapi-python==0.2.6
phonenumberslite==8.12.28
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
